# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
security:
  - bearerAuth: []
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/ApiError"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/ApiError"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/ApiError"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/ApiError"
...
  /sessions/{sessionId}/generate/epics:
    post:
      operationId: generateEpics
      tags: [sessions]
      summary: Generate epics
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateWorkflowStepRequest"
      responses:
        "200":
          description: Updated workflow session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowSession"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/ApiError"
  /sessions/{sessionId}/generate/stories:
    post:
      operationId: generateStories
      tags: [sessions]
      summary: Generate stories
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateWorkflowStepRequest"
      responses:
        "200":
          description: Updated workflow session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowSession"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/ApiError"
...
  /integrations/config:
    get:
      operationId: listIntegrationConfigs
      tags: [integrations]
      summary: List integration configurations
      responses:
        "200":
          description: Integration configuration list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationConfigListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/ApiError"
  /integrations/config/{type}:
    put:
      operationId: updateIntegrationConfig
      tags: [integrations]
      summary: Update integration configuration
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIntegrationConfigRequest"
      responses:
        "200":
          description: Integration configuration updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationConfig"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/ApiError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
  responses:
    ApiError:
      description: API error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Forbidden:
      description: Authenticated but not authorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    ApiError:
      type: object
      properties:
        message:
          type: string
      required: [message]
    Phase:
      type: string
      enum:
        - intake
        - clarification
        - prd
        - epics
        - stories
        - quality
        - devReview
        - export
    PhaseStatus:
      type: string
      enum:
        - not-started
        - in-progress
        - complete
        - needs-attention
    ReviewStatus:
      type: string
      enum:
        - pending
        - approved
        - needs-clarification
        - too-large
        - technically-risky
        - blocked
        - missing-ac
    EpicPriority:
      type: string
      enum: [P0, P1, P2]
    StoryPriority:
      type: string
      enum: [P0, P1, P2, P3]
    ReadinessLabel:
      type: string
      enum:
        - Ready for Jira
        - Minor review needed
        - Needs PM refinement
        - Not ready
    WarningSeverity:
      type: string
      enum: [error, warning, info]
...
    PmRevisionStatus:
      type: string
      enum:
        - not-started
        - in-progress
        - resolved
    GenerationMode:
      type: string
      enum: [demo, live, unavailable]
    GenerationStatus:
      type: string
      enum: [idle, running, succeeded, failed, unavailable]
    GenerationStepState:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/GenerationStatus"
        mode:
          $ref: "#/components/schemas/GenerationMode"
        promptVersion:
          type: string
        updatedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        errorMessage:
          oneOf:
            - type: string
            - type: "null"
      required: [status, mode, promptVersion, updatedAt, errorMessage]
    WorkflowGenerationState:
      type: object
      properties:
        clarification:
          $ref: "#/components/schemas/GenerationStepState"
        prd:
          $ref: "#/components/schemas/GenerationStepState"
        epics:
          $ref: "#/components/schemas/GenerationStepState"
        stories:
          $ref: "#/components/schemas/GenerationStepState"
        quality:
          $ref: "#/components/schemas/GenerationStepState"
      required: [clarification, prd, epics, stories, quality]
    PhasesRecord:
      type: object
      additionalProperties:
        $ref: "#/components/schemas/PhaseStatus"
      required:
        - intake
        - clarification
        - prd
        - epics
        - stories
        - quality
        - devReview
        - export
      properties:
        intake:
          $ref: "#/components/schemas/PhaseStatus"
        clarification:
          $ref: "#/components/schemas/PhaseStatus"
        prd:
          $ref: "#/components/schemas/PhaseStatus"
        epics:
          $ref: "#/components/schemas/PhaseStatus"
        stories:
          $ref: "#/components/schemas/PhaseStatus"
        quality:
          $ref: "#/components/schemas/PhaseStatus"
        devReview:
          $ref: "#/components/schemas/PhaseStatus"
        export:
          $ref: "#/components/schemas/PhaseStatus"
    ClarificationQuestion:
      type: object
      properties:
        id:
          type: string
        group:
          type: string
        text:
          type: string
        required:
          type: boolean
        answer:
          type: string
        skipped:
          type: boolean
      required: [id, group, text, required, answer, skipped]
    PrdSection:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        complete:
          type: boolean
        order:
          type: integer
      required: [id, title, content, complete, order]
    ReadinessScore:
      type: object
      properties:
        total:
          type: integer
        clarity:
          type: integer
        acceptanceCriteria:
          type: integer
        businessAlignment:
          type: integer
        technicalFeasibility:
          type: integer
        testability:
          type: integer
        edgeCasesErrorHandling:
          type: integer
        dependenciesDesignLocalization:
          type: integer
        label:
          $ref: "#/components/schemas/ReadinessLabel"
      required:
        - total
        - clarity
        - acceptanceCriteria
        - businessAlignment
        - technicalFeasibility
        - testability
        - edgeCasesErrorHandling
        - dependenciesDesignLocalization
        - label
    QualityWarning:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        message:
          type: string
        severity:
          $ref: "#/components/schemas/WarningSeverity"
      required: [id, type, message, severity]
    DeveloperReview:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ReviewStatus"
        comment:
          type: string
        reviewerName:
          type: string
        timestamp:
          type: string
          format: date-time
        pmRevisionStatus:
          $ref: "#/components/schemas/PmRevisionStatus"
      required: [status, comment, reviewerName, timestamp, pmRevisionStatus]
    Epic:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        title:
          type: string
        businessObjective:
          type: string
        scopeSummary:
          type: string
        prdRequirements:
          type: array
          items:
            type: string
        priority:
          $ref: "#/components/schemas/EpicPriority"
        dependencies:
          type: array
          items:
            type: string
        risks:
          type: array
          items:
            type: string
        jiraEpicDescription:
          type: string
        storyCount:
          type: integer
      required:
        - id
        - sessionId
        - title
        - businessObjective
        - scopeSummary
        - prdRequirements
        - priority
        - dependencies
        - risks
        - jiraEpicDescription
        - storyCount
    Story:
      type: object
      properties:
        id:
          type: string
        epicId:
          type: string
        sessionId:
          type: string
        title:
          type: string
        userStory:
          type: string
        description:
          type: string
        acceptanceCriteria:
          type: array
          items:
            type: string
        priority:
          $ref: "#/components/schemas/StoryPriority"
        labels:
          type: array
          items:
            type: string
        components:
          type: array
          items:
            type: string
        dependencies:
          type: array
          items:
            type: string
        edgeCases:
          type: array
          items:
            type: string
        errorHandling:
          type: string
        localizationNotes:
          type: string
        designNotes:
          type: string
        analyticsNotes:
          type: string
        qaNotes:
          type: string
        technicalNotes:
          type: string
        openQuestions:
          type: array
          items:
            type: string
        readinessScore:
          $ref: "#/components/schemas/ReadinessScore"
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/QualityWarning"
        reviewStatus:
          $ref: "#/components/schemas/ReviewStatus"
        developerReview:
          oneOf:
            - $ref: "#/components/schemas/DeveloperReview"
            - type: "null"
      required:
        - id
        - epicId
        - sessionId
        - title
        - userStory
        - description
        - acceptanceCriteria
        - priority
        - labels
        - components
        - dependencies
        - edgeCases
        - errorHandling
        - localizationNotes
        - designNotes
        - analyticsNotes
        - qaNotes
        - technicalNotes
        - openQuestions
        - readinessScore
        - warnings
        - reviewStatus
...
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        jiraKey:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, jiraKey, createdAt, updatedAt]
    WorkflowSession:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        name:
          type: string
        inputType:
          type: string
        outputDepth:
          type: string
        jiraKey:
          type: string
        targetUsers:
          type: array
          items:
            type: string
        businessGoal:
          type: string
        knownConstraints:
          type: string
        labels:
          type: array
          items:
            type: string
        rawInput:
          type: string
        currentPhase:
          $ref: "#/components/schemas/Phase"
        phases:
          $ref: "#/components/schemas/PhasesRecord"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        clarificationQuestions:
          type: array
          items:
            $ref: "#/components/schemas/ClarificationQuestion"
        prdSections:
          type: array
          items:
            $ref: "#/components/schemas/PrdSection"
        epics:
          type: array
          items:
            $ref: "#/components/schemas/Epic"
        stories:
          type: array
          items:
            $ref: "#/components/schemas/Story"
        generation:
          $ref: "#/components/schemas/WorkflowGenerationState"
      required:
        - id
        - projectId
        - name
        - inputType
        - outputDepth
        - jiraKey
        - targetUsers
        - businessGoal
        - knownConstraints
        - labels
        - rawInput
        - currentPhase
        - phases
        - createdAt
        - updatedAt
        - clarificationQuestions
        - prdSections
        - epics
        - stories
        - generation
...
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        jiraKey:
          type: string
      required: [name]
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        jiraKey:
          type: string
      additionalProperties: false
...
    GenerateWorkflowStepRequest:
      type: object
      properties:
        force:
          type: boolean
      additionalProperties: false
...
    ProjectListResponse:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
      required: [projects]
...
    IntegrationConfig:
      type: object
      properties:
        id:
          type: string
        integrationType:
          type: string
        enabled:
          type: boolean
        configured:
          type: boolean
      required: [id, integrationType, enabled, configured]
    IntegrationConfigListResponse:
      type: object
      properties:
        integrations:
          type: array
          items:
            $ref: "#/components/schemas/IntegrationConfig"
      required: [integrations]
    UpdateIntegrationConfigRequest:
      type: object
      properties:
        enabled:
          type: boolean
        config:
          type: object
          additionalProperties:
            type: string
      required: [enabled]