# lib/api-zod/src/generated/types/integrationConfigListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { IntegrationConfig } from "./integrationConfig";

export interface IntegrationConfigListResponse {
  integrations: IntegrationConfig[];
}