# lib/api-zod/src/generated/types/updateIntegrationConfigRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateIntegrationConfigRequestConfig } from "./updateIntegrationConfigRequestConfig";

export interface UpdateIntegrationConfigRequest {
  enabled: boolean;
  config?: UpdateIntegrationConfigRequestConfig;
}